+++ lib/open890_web/components/buttons.ex
@@ -24,6 +24,52 @@ defmodule Open890Web.Components.Buttons do
     """
   end
 
+  def band_buttons(assigns) do
+    ~H"""
+      <div class="ui grid">
+        <div class="row">
+          <div class="four wide column left aligned">
+
+            <div id="modeKeys" class="ui vertical big buttons">
+              <.cmd_button tabindex="0" classes="regular" cmd="MK0">LSB/USB</.cmd_button>
+              <.cmd_button tabindex="0" classes="regular" cmd="MK1">CW/CW-R</.cmd_button>
+              <.cmd_button tabindex="0" classes="regular" cmd="MK2">FSK/PSK</.cmd_button>
+              <.cmd_button tabindex="0" classes="regular" cmd="MK3">FM/AM</.cmd_button>
+              <.cmd_button tabindex="0" classes="regular" cmd="MK4">FSK/FSK-R</.cmd_button>
+              <.cmd_button tabindex="0" classes="regular" cmd="MK5">PSK/PSK-R</.cmd_button>
+            </div>
+
+          </div>
+          <div class="twelve wide column">
+            <div class="ui equal width grid">
+              <div class="row">
+                <div class="column"><.cmd_button cmd="BU000" classes="big regular" tabindex="0" fluid>1.8</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU001" classes="big regular" tabindex="0" fluid>3.5</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU002" classes="big regular" tabindex="0" fluid>7</.cmd_button></div>
+              </div>
+              <div class="row">
+                <div class="column"><.cmd_button cmd="BU003" classes="big regular" tabindex="0" fluid>10</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU004" classes="big regular" tabindex="0" fluid>14</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU005" classes="big regular" tabindex="0" fluid>18</.cmd_button></div>
+              </div>
+              <div class="row">
+                <div class="column"><.cmd_button cmd="BU006" classes="big regular" tabindex="0" fluid>21</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU007" classes="big regular" tabindex="0" fluid>24</.cmd_button></div>
+                <div class="column"><.cmd_button cmd="BU008" classes="big regular" tabindex="0" fluid>28</.cmd_button></div>
+              </div>
+              <div class="row">
+                <div class="column"><.cmd_button cmd="BU010" classes="big regular" tabindex="0" fluid>GEN</.cmd_button></div>
+                <div class="column"></div>
+                <div class="column"><.cmd_button cmd="BU009" classes="big regular" tabindex="0" fluid>50</.cmd_button></div>
+              </div>
+            </div>
+
+          </div>
+        </div>
+      </div>
+    """
+  end
+
   def send_button(assigns) do
     tx_state = assigns.value
 
@@ -428,6 +474,7 @@ defmodule Open890Web.Components.Buttons do
     """
   end
 
+  # attr :tabindex, :number, required: false, default: 0
   def cmd_button(assigns) do
     fluid_class =
       assigns
@@ -451,7 +498,15 @@ defmodule Open890Web.Components.Buttons do
         "small"
       end
 
-    classes = "ui #{size_class} black button #{assigned_classes} #{fluid_class}"
+    color_class =
+      if ~w(primary secondary regular)
+        |> Enum.any?(fn size -> assigned_classes |> String.contains?(size) end) do
+          ""
+        else
+          "black"
+        end
+
+    classes = "ui #{size_class} #{color_class} button #{assigned_classes} #{fluid_class}"
 
     ~H"""
       <div class={classes} phx-click="cmd" phx-value-cmd={@cmd}>